requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
pyngrok==7.2.3
httpx==0.24.1
anthropic==0.25.0
